refactor (consolidate testing samples under "/test.ui" folder)
