--- code/system.ui/react.dev/src/test.sample/index.ts
@@ -1,2 +1,0 @@
-export * from './specs';
-export * from './sample.DevTools';
