+++ code/system.ui/react.dev/src/test.ui/entry.Specs.mts
@@ -1,6 +1,6 @@
 export const SampleSpecs = {
-  'sample.MySample': () => import('../test.sample/specs/MySample.SPEC'),
-  'sample.DevTools.Button': () => import('../test.sample/sample.DevTools/s.ui.Button.SPEC'),
+  'sample.MySample': () => import('../test.ui/sample.specs/MySample.SPEC'),
+  'sample.DevTools.Button': () => import('../test.ui/sample.DevTools/s.ui.Button.SPEC'),
 };
 
 export const SelfDevSpecs = {
